fix migration for non-existent table

+++ db/migrate/20170619171501_drop_test.rb
@@ -1,5 +1,10 @@
 class DropTest < ActiveRecord::Migration[5.0]
-  def change
-    drop_table :test
+  def up
+    if ActiveRecord::Base.connection.table_exists? :test
+      drop_table :test
+    end
+  end
+
+  def down
   end
 end
